Refactor template generation code out of the KIDL parser
Doesn't belong there - the KIDL parser should parse KIDL and nothing
more. The template generation code has been moved into a KIDL Visitor
class in the same package as the `TemplateBasedGenerator`.

The code is mostly C&P other than changes to allow for the fact the code
no longer has access to instance variables and use the visitor pattern
where possible (which isn't much w/o making drastic changes).

Tested by generating the template structure with the code in the visitor
and the code in the `forTemplates` methods before deleting said methods,
and throwing an exception if they weren't `equals()`.
