+++ src/java/us/kbase/kidl/KbTypedef.java
@@ -1,15 +1,7 @@
 package us.kbase.kidl;
 
-import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.LinkedHashMap;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.Map;
-import java.util.Set;
-
-//TODO CODE don't make a dependency on SDK compiled code in the SDK
-import us.kbase.common.service.Tuple2;
 
 /**
  * Class represents type definition (or named type) in spec-file.
@@ -156,27 +148,6 @@ public class KbTypedef implements KbType, KbModuleDef {
 	@Override
 	public void afterCreation() {
 		// Default implementation - just do nothing
-	}
-
-	@Override
-	public Map<String, Object> forTemplates() {
-		Map<String, Object> ret = new LinkedHashMap<String, Object>();
-		ret.put("name", getName());
-		ret.put("comment", getComment());
-		ret.put("english", getTypeInEnglish(getAliasType()));
-		return ret;
-	}
-
-	private static String getTypeInEnglish(KbType type) {
-		Set<String> allKeys = new HashSet<String>();
-		List<String> additional = new ArrayList<>();
-		LinkedList<Tuple2<String, KbType>> subQueue = new LinkedList<Tuple2<String, KbType>>();
-		StringBuilder ret = new StringBuilder(Utils.getEnglishTypeDescr(type, subQueue, allKeys, additional));
-		if (additional.size() > 0)
-			ret.append(":\n");
-		for (String add : additional)
-			ret.append(add).append("\n");
-		return ret.toString();
 	}
 
 	@Override
